+++ pkg/anonymity/doc.go
@@ -1,5 +1,0 @@
-// Package anonymity allows you to perform network operations on top of anonymizing traffic.
-//
-// The package basically uses the fifth^ stage of anonymity with a queue-based problem.
-// All applied connections use friend-to-friend (F2F) communications.
-package anonymity