+++ pkg/anonymity/qb/action.go
@@ -1,4 +1,4 @@
-package anonymity
+package qb
 
 const (
 	cAction32bitMask = sAction(1 << 31)