+++ pkg/anonymity/qb/anonymity_test.go
@@ -1,5 +1,5 @@
 // nolint: err113
-package anonymity
+package qb
 
 import (
 	"context"
@@ -11,8 +11,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/number571/go-peer/pkg/anonymity/adapters"
-	"github.com/number571/go-peer/pkg/anonymity/queue"
+	"github.com/number571/go-peer/pkg/anonymity/qb/adapters"
+	"github.com/number571/go-peer/pkg/anonymity/qb/queue"
 	"github.com/number571/go-peer/pkg/client"
 	"github.com/number571/go-peer/pkg/crypto/asymmetric"
 	"github.com/number571/go-peer/pkg/crypto/hashing"
@@ -26,7 +26,7 @@ import (
 	"github.com/number571/go-peer/pkg/storage/database"
 	testutils "github.com/number571/go-peer/test/utils"
 
-	anon_logger "github.com/number571/go-peer/pkg/anonymity/logger"
+	anon_logger "github.com/number571/go-peer/pkg/anonymity/qb/logger"
 	"github.com/number571/go-peer/pkg/network/conn"
 )
 
@@ -189,7 +189,7 @@ func TestFetchPayload(t *testing.T) {
 
 	nodes[1].HandleFunc(
 		tcHead,
-		func(_ context.Context, _ INode, _ asymmetric.IPubKey, reqBytes []byte) ([]byte, error) {
+		func(_ context.Context, _ INode, _ uint64, _ asymmetric.IPubKey, reqBytes []byte) ([]byte, error) {
 			return []byte(fmt.Sprintf("echo: '%s'", string(reqBytes))), nil
 		},
 	)
@@ -238,7 +238,7 @@ func TestBroadcastPayload(t *testing.T) {
 	chResult := make(chan string)
 	nodes[1].HandleFunc(
 		tcHead,
-		func(_ context.Context, _ INode, _ asymmetric.IPubKey, reqBytes []byte) ([]byte, error) {
+		func(_ context.Context, _ INode, _ uint64, _ asymmetric.IPubKey, reqBytes []byte) ([]byte, error) {
 			res := fmt.Sprintf("echo: '%s'", string(reqBytes))
 			go func() { chResult <- res }()
 			return nil, nil
@@ -386,7 +386,7 @@ func TestHandleWrapper(t *testing.T) {
 
 	node.HandleFunc(
 		111,
-		func(_ context.Context, _ INode, _ asymmetric.IPubKey, _ []byte) ([]byte, error) {
+		func(_ context.Context, _ INode, _ uint64, _ asymmetric.IPubKey, _ []byte) ([]byte, error) {
 			return nil, errors.New("some error") //nolint:err113
 		},
 	)
@@ -600,7 +600,7 @@ func testRunNodes(ctx context.Context, t *testing.T, timeWait time.Duration, add
 	for _, node := range nodes {
 		node.HandleFunc(
 			tcHead,
-			func(_ context.Context, _ INode, _ asymmetric.IPubKey, reqBytes []byte) ([]byte, error) {
+			func(_ context.Context, _ INode, _ uint64, _ asymmetric.IPubKey, reqBytes []byte) ([]byte, error) {
 				// send response
 				return []byte(string(reqBytes) + " (response)"), nil
 			},