+++ pkg/anonymity/qb/doc.go
@@ -1,0 +1,5 @@
+// Package qb allows you to perform network operations on top of anonymizing traffic.
+//
+// The package basically uses the fifth^ stage of qb with a queue-based problem.
+// All applied connections use friend-to-friend (F2F) communications.
+package qb