+++ pkg/anonymity/qb/head.go
@@ -1,4 +1,4 @@
-package anonymity
+package qb
 
 type iHead interface {
 	uint64() uint64