+++ pkg/anonymity/qb/settings.go
@@ -1,4 +1,4 @@
-package anonymity
+package qb
 
 import (
 	"time"